[prowlarr] add options to automatically sync settings
These use a bunch of `jq` (as well as the new api-caller utilities) to
automatically add or upate apps, indexers, and tag labels within
Prowlarr.

For apps and indexers, the general strategy is to either find an
existing item, or use the "schema" (provided by the API) to supply a
default/base value; in either case, we then apply the NixOS config to
that value and then add or overwrite the item.